+++ app/src/main/java/com/example/roomie/components/chat/AttachmentPreviewSection.kt
@@ -9,8 +9,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 
-
-
 @Composable
 fun AttachmentPreviewSection(
     attachedFiles: List<AttachedFile>,
