--- app/src/main/java/com/example/roomie/components/chat/AttachmentUtilities.kt
@@ -1,10 +1,0 @@
-package com.example.roomie.components.chat
-
-import android.net.Uri
-
-data class AttachedFile(
-    val uri: Uri,
-    val name: String,
-    val type: String, // "image", "video", "pdf", etc.
-    val size: Long? = null
-)
