+++ app/src/main/java/com/example/roomie/screens/SingleChatScreen.kt
@@ -43,6 +43,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
+
 import com.example.roomie.components.chat.AttachedFile
 import com.example.roomie.components.chat.ChatManager
 import com.example.roomie.components.chat.Message
